Rename loadSpots to load
